+++ physprep/utils.py
@@ -1,0 +1,31 @@
+"""Utilities for Physprep"""
+
+import json
+import pickle
+
+
+def load_json(filename):
+    """
+    Parameters
+    ----------
+    filename : str
+        File path of the .json to load.
+
+    Returns
+    -------
+    data : dict
+        Dictionary with the content of the .json passed in argument.
+    """
+    try:
+        with open(filename, "r") as tmp:
+            data = json.loads(tmp.read())
+        tmp.close()
+    except Exception:
+        try:
+            with open(filename, "rb") as tmp:
+                data = pickle.load(tmp)
+            tmp.close()
+        except Exception:
+            pass
+
+    return data
